+++ program.py
@@ -1,6 +1,6 @@
-from Scripts.Git.Git import Git
+from Scripts.Git import Git
+from Scripts.Packages import PackageCollection
 from Scripts.DriveMounting import DriveCollection
-from Scripts.Packages.Package import PackageCollection
 from Scripts.ServerConfig import ServerConfig
 from Scripts.FolderSyncing import FolderSyncing
 from Scripts.SystemLinks import SymLinks
